Changed weapon/projectile structure. projectiles now in same filew as their weapons.
