+++ src/config.json
@@ -68,19 +68,10 @@
                 "projectile_gravity": 400.0,
                 "projectile_draw_size_radius": 5
             },
-            "basic_cannon": {
-                "explosion_radius": 30,
-                "center_damage": 50,
+            "small_bomb": {
+                "explosion_radius": 25,
+                "center_damage": 40,
                 "power_to_velocity_scale": 5.0
-            }
-        },
-        "projectiles": {
-            "defaults": {
-                "gravity": 400.0,
-                "draw_size_radius": 5,
-                "explosion_radius": 30,
-                "max_player_damage": 50,
-                "terrain_impact_strength": 100
             }
         },
         "game_settings": {
